# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Rokwire Rokwire Building Block API Documentation.",
        "title": "Rokwire Gatewauy Building Block API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "localhost",
    "basePath": "/gateway/api",
    "paths": {
...
        "/laundry/initrequest": {
            "get": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Returns the problem codes and pending service reqeust status for a laundry machine.",
                "operationId": "InitRequest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "machine service tag id",
                        "name": "machineid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MachineRequestDetail"
                        }
                    }
                }
            }
        },
        "/laundry/requestservice": {
            "post": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "RequestService",
                "parameters": [
                    {
                        "description": "body json",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ServiceSubmission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ServiceRequestResult"
                        }
                    }
                }
            }
        },
        "/laundry/roomdetail": {
            "get": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Returns the list of machines and the number of washers and dryers available in a laundry room",
                "operationId": "Room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RoomDetail"
                        }
                    }
                }
            }
        },
        "/laundry/rooms": {
            "get": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Get list of all campus laundry rooms",
                "operationId": "Rooms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Organization"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Appliance": {
            "type": "object",
            "properties": {
                "applianceType": {
                    "type": "string"
                },
                "averageCycleTime": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timeRemaining": {
                    "type": "integer"
                }
            }
        },
...
        "model.LaundryDetails": {
            "type": "object",
            "properties": {
                "floor": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "model.LaundryRoom": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "location": {
                    "$ref": "#/definitions/model.LaundryDetails"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.MachineRequestDetail": {
            "type": "object",
            "properties": {
                "machineID": {
                    "type": "string"
                },
                "machineType": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "openIssue": {
                    "type": "boolean"
                },
                "problemCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.Organization": {
            "type": "object",
            "properties": {
                "laundryRooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.LaundryRoom"
                    }
                },
                "schoolName": {
                    "type": "string"
                }
            }
        },
...
        "model.RoomDetail": {
            "type": "object",
            "properties": {
                "appliances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Appliance"
                    }
                },
                "campusName": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/model.LaundryDetails"
                },
                "numDryers": {
                    "type": "integer"
                },
                "numWashers": {
                    "type": "integer"
                },
                "roomName": {
                    "type": "string"
                }
            }
        },
        "model.ServiceRequestResult": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "requestNumber": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.ServiceSubmission": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "machineid": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "problemcode": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminUserAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header (add admin id token with Bearer prefix to the Authorization value)"
        },
        "InternalAuth": {
            "type": "apiKey",
            "name": "INTERNAL-API-KEY",
            "in": "header"
        },
        "RokwireAuth": {
            "type": "apiKey",
            "name": "ROKWIRE-API-KEY",
            "in": "header"
        },
        "UserAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header (add client id token with Bearer prefix to the Authorization value)"
        }
    }
}